+++ src/1_input_parsing/a_file_reading/set_values.c
@@ -3,18 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   set_values.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jkauker <jkauker@student.42heilbronn.de    +#+  +:+       +#+        */
+/*   By: fschuber <fschuber@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/28 08:12:38 by jkauker           #+#    #+#             */
-/*   Updated: 2024/06/10 12:18:49 by jkauker          ###   ########.fr       */
+/*   Updated: 2024/06/14 10:54:44 by fschuber         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../../include/cub3d.h"
-
-int		split_len(char **split);
-bool	free_split(char	**split, bool ret);
-bool	regex(char *line, char *reg);
 
 bool	set_player_spawn(char dir, t_vec2 pos, t_tile_type ***map)
 {
@@ -41,6 +37,20 @@ bool	set_player_spawn(char dir, t_vec2 pos, t_tile_type ***map)
 	player()->spawn_transform.pos = pos;
 	player()->transform = player()->spawn_transform;
 	*(map[(int)pos.y][(int)pos.x]) = FLOOR;
+	return (true);
+}
+
+bool	set_goal(t_vec2 pos, t_tile_type ***map)
+{
+	static int	goal_count = 0;
+
+	if (++goal_count > MAX_GOAL_COUNT)
+	{
+		logger(LOGGER_WARNING, "Maximum goal count already reached!");
+		return (false);
+	}
+	*(map[(int)pos.y][(int)pos.x]) = FLOOR;
+	create_entity(pos, (t_vec2){0, 1}, GOAL);
 	return (true);
 }